Remove tailwindcss     - Sass configuration     - Styles addition (normalize, variables and mixins)     - Create FrontController, front templates     - Navbar.jsx modifications:         - FrontNav.jsx and BackNav.jsx added         - Create front_base.html.twig for twig files extension.
