+++ assets/styles/app.scss
@@ -1,0 +1,111 @@
+@import "../styles/mixins.scss";
+@import "../styles/variables.scss";
+
+// _________________||| HEADER \\\________________
+
+// _________________||| NAVBAR \\\________________
+
+
+.frontNav {
+    @include dFlex(row);
+    @include centerXY(space-between, center);
+    background: linear-gradient( 100deg, $primaryLight, $primaryDark);
+    color: $primaryDark;
+    padding-left: 20px;
+    box-shadow: 0 0 4px 1px $primaryDark;
+
+
+    // _________________||| LOGO \\\________________
+    .logo {
+        font-size: 22px;
+        width: 10vw;
+        transition: color 0.3s ease; 
+        cursor: pointer;
+        h2 {
+            color: white;
+            text-shadow: 0 0 7px black;
+            margin: 0
+        }
+    }
+    
+    // _________________||| LINKS \\\________________
+    .navLinks{
+        @include dFlex(row);
+        @include centerXY(center, center);
+        font-size: 18px;
+        width: 75vw;
+        padding: 0;
+        li {
+            list-style: none;
+            width: 110px;
+            margin-inline: 20px;            
+            a {
+                display: block;
+                width: 100%;
+                @include roundedBtn;
+                color: white; 
+                text-decoration: none; 
+                background-color: $primaryDark;
+                opacity: 0.8;
+                text-align: center;
+                &:hover {
+                    color: $primaryDark;
+                    cursor: pointer;
+                    box-shadow: 0 0 5px black ;
+                    background-color: $primaryLight;
+                }
+            }
+        }
+    }
+    .loginBtn {
+        font-size: 20px;
+        border-left: solid 1px white;
+        width: 10vw;
+        height: 70px;
+        padding: auto 0;
+        a{
+            display: block;            
+            text-align: center;
+            text-decoration: none;
+            color: white;
+            width: 100%;
+            height: 100%;
+            padding-top: 18%;
+        }
+        &:hover{
+            background-color: $primaryLight;
+            cursor: pointer;
+            a{
+                display: block;
+                color: $primaryDark;
+            }
+        }
+    }
+}
+
+// _________________||| TITLE & DESCRIPTION \\\________________
+
+header {
+    background-image: url("../images/library.jpg");    
+    background-size: cover;
+    @include dFlex(column);
+    @include centerXY(center, center);
+    section {
+        @include dFlex(column);
+        @include centerXY(center, center);
+        width: 55vw;
+        margin-block: 50px;
+        background-color: $lightGrey;
+        opacity: 0.8;
+        border-radius: 15px;
+        text-align: center;
+        h1 {
+            margin-inline: 20px;
+        }
+        .description{
+            padding-inline: 30px;
+            font-size: 1.4rem;
+            line-height: 2.5rem;
+        }
+    }
+}
