+++ assets/styles/variables.scss
@@ -1,0 +1,6 @@
+// Visual Identity
+
+$primaryBasic: #239962;
+$primaryLight: #a3f5d3;
+$primaryDark: #0f422d;
+$lightGrey: #d4d4d4;
