+++ package.json
@@ -7,14 +7,12 @@
         "@symfony/stimulus-bridge": "^3.2.0",
         "@symfony/ux-react": "file:vendor/symfony/ux-react/assets",
         "@symfony/webpack-encore": "^4.0.0",
-        "autoprefixer": "^10.4.16",
         "core-js": "^3.23.0",
-        "postcss": "^8.4.31",
-        "postcss-loader": "^7.3.3",
         "react": "^18.0",
         "react-dom": "^18.0",
         "regenerator-runtime": "^0.13.9",
-        "tailwindcss": "^3.3.3",
+        "sass": "^1.69.5",
+        "sass-loader": "^13.3.2",
         "webpack": "^5.74.0",
         "webpack-cli": "^4.10.0",
         "webpack-notifier": "^1.15.0"
@@ -28,8 +26,8 @@
         "build": "encore production --progress"
     },
     "dependencies": {
-        "@headlessui/react": "^1.7.17",
-        "@heroicons/react": "^2.0.18",
+        "autoprefixer": "^10.4.16",
+        "postcss": "^8.4.31",
         "react-router-dom": "^6.17.0"
     }
 }
