+++ src/Controller/FrontController.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Controller;
+
+use App\Entity\User;
+use App\Repository\UserRepository;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+
+#[Route('/')]
+class FrontController extends AbstractController
+{
+    #[Route('/', name: 'home')]
+    public function home(): Response
+    {
+        
+        return $this->render('front/home.html.twig');
+    }
+
+}
