--- tailwind.config.js
@@ -1,13 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./assets/**/*.js",
-    "./templates/**/*.html.twig",
-    "./assets/react/controllers/*.jsx",
-],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
