+++ templates/front/home.html.twig
@@ -1,0 +1,19 @@
+{% extends 'front_base.html.twig' %}
+
+{% block title %}Home{% endblock %}
+
+{% block body %}
+    <header>
+        <section>
+            <h1>Welcome to Compass Education limited, the perfect place to learn a new language!</h1>
+            <p class="description">Lorem ipsum dolor sit, amet consectetur adipisicing elit. Molestias, reiciendis incidunt. Laborum asperiores, eveniet, quas deleniti delectus nisi aliquid cum architecto molestias laudantium corporis quos.</p>
+        </section>
+    </header>
+    <main>
+        <section></section>
+        <section></section>
+    </main>
+    <footer>
+    </footer>
+    
+{% endblock %}
