+++ templates/front_base.html.twig
@@ -8,7 +8,7 @@
     {% endblock %}
     </head>
     {% block head %}
-        <div {{ react_component('BackNav') }}></div> 
+        <nav {{ react_component('FrontNav') }}></nav> 
     {% endblock %}
     {% block body %}
     {% endblock %}
