+++ templates/test/flag.html.twig
@@ -43,4 +43,6 @@ ________________________________________________________________________________
     <p>image : {{ newFlag.image }}</p>
     <p>Language: {{ newFlag.language.name }}</p>
 
+    
+
 {% endblock %}
